Add ability for connection to fail after set number of retries
Orginal libetrv behaviour was to retry forever in eTRVDevice.connect().

This is not good for real world applications where devices run out of battery
or go out of BLE range due to interference. This commit adds new constructor property
retry_limit for eTRVDevice that allows to throw exception to user application after set number of failures.